+++ test/case-addmap.sh
@@ -1,3 +1,5 @@
 
-./xsvutils ./test/data/sample-4.tsv addmap code-name code ./test/data/prefecture.tsv > ./var/test/case-addmap-1.tsv
+./xsvutils ./test/data/sample-4.tsv addmap code-name code ./test/data/prefecture.tsv > ./var/test/addmap-1.tsv
+
+./xsvutils -v2 ./test/data/sample-4.tsv insmap code code-name ./test/data/prefecture.tsv > ./var/test/addmap-2.tsv