+++ test/case-diff.sh
@@ -1,0 +1,3 @@
+
+./xsvutils -v2 ./test/data/sample-4.tsv diff ./test/data/sample-5.tsv > ./var/test/diff.txt
+