+++ test/case-number.sh
@@ -1,3 +1,3 @@
 
-./xsvutils -v2 ./test/data/sample-number.tsv groupsum                 > ./var/test/case-number-1.tsv
+./xsvutils -v2 ./test/data/sample-number.tsv groupsum -v2 > ./var/test/case-number-1.tsv