+++ test/test.sh
@@ -9,6 +9,7 @@ bash ./test/case-newline.sh
 bash ./test/case-ltsv.sh
 
 bash ./test/case-range.sh
+bash ./test/case-where.sh
 bash ./test/case-uriparams.sh
 bash ./test/case-facetcount.sh
 bash ./test/case-treetable.sh
@@ -17,6 +18,7 @@ bash ./test/case-wordsflags.sh
 bash ./test/case-addmap.sh
 bash ./test/case-sort.sh
 bash ./test/case-join.sh
+bash ./test/case-diff.sh
 bash ./test/case-datetime.sh
 bash ./test/case-number.sh